Feat(report): PnLレポート生成を独立したマイクロサービスに分離
これまで `datastore` に混在していたPnLレポートの生成ロジックを、独立した `report-generator` マイクロサービスに分離しました。

主な変更点：
- `cmd/report/main.go`: 新しいレポート生成サービスのエントリーポイント。定期的に `trades` テーブルからデータを集計し、`pnl_reports` テーブルに保存します。
- `internal/report`: レポートの分析・保存ロジックを格納する新しいパッケージ。
- `internal/datastore`: レポート生成関連の責務を `internal/report` に移譲し、関連コードを削除。
- `docker-compose.yml`: `report-generator` サービスを定義し、独立したコンテナとして実行されるように設定。
- `cmd/report/Dockerfile`: `report-generator` サービスをビルドするためのDockerfile。

この変更により、関心の分離が促進され、各サービスの責務が明確になりました。メンテナンス性とスケーラビリティが向上し、今後の機能追加や修正が容易になります。
